+++ php/tests/Application/Command/ClassInfoTest/CompoundClassPropertyStatement.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace A;
+
+class TestClass
+{
+    /**
+     * This is the summary.
+     *
+     * This is a long description.
+     *
+     * @var Foo1 $testProperty1 A description of the first property.
+     * @var Foo2 $testProperty2 A description of the second property.
+     */
+    protected $testProperty1, $testProperty2;
+}